refactor: split auth session lifecycle helpers
